[move] 패키지 구조 및 파일 위치 변경
