+++ src/main/java/com/umc/ttt/domain/bookClub/service/BookClubQueryServiceImpl.java
@@ -9,7 +9,7 @@ import com.umc.ttt.domain.bookClub.dto.BookClubResponseDTO;
 import com.umc.ttt.domain.bookClub.entity.BookClub;
 import com.umc.ttt.domain.bookClub.entity.BookClubMember;
 import com.umc.ttt.domain.bookClub.entity.ReadingRecord;
-import com.umc.ttt.domain.bookClub.handler.BookClubHandler;
+import com.umc.ttt.global.apiPayload.exception.handler.BookClubHandler;
 import com.umc.ttt.domain.bookClub.repository.BookClubMemberRepository;
 import com.umc.ttt.domain.bookClub.repository.BookClubRepository;
 import com.umc.ttt.domain.bookClub.repository.ReadingRecordRepository;
@@ -18,8 +18,8 @@ import com.umc.ttt.domain.member.dto.MemberResponseDTO;
 import com.umc.ttt.domain.member.entity.Member;
 import com.umc.ttt.domain.scrap.repository.BookScrapRepository;
 import com.umc.ttt.global.apiPayload.code.status.ErrorStatus;
-import com.umc.ttt.home.converter.HomeConverter;
-import com.umc.ttt.home.dto.HomeResponseDTO;
+import com.umc.ttt.domain.home.converter.HomeConverter;
+import com.umc.ttt.domain.home.dto.HomeResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
