+++ src/main/java/com/umc/ttt/domain/bookClub/service/BookClubServiceImpl.java
@@ -6,13 +6,13 @@ import com.umc.ttt.domain.bookClub.dto.BookClubResponseDTO;
 import com.umc.ttt.domain.bookClub.entity.BookClub;
 import com.umc.ttt.domain.bookClub.entity.BookClubMember;
 import com.umc.ttt.domain.bookClub.entity.ReadingRecord;
-import com.umc.ttt.domain.bookClub.handler.BookClubHandler;
+import com.umc.ttt.global.apiPayload.exception.handler.BookClubHandler;
 import com.umc.ttt.domain.bookClub.repository.BookClubMemberRepository;
 import com.umc.ttt.domain.bookClub.repository.BookClubRepository;
 import com.umc.ttt.domain.bookClub.repository.ReadingRecordRepository;
-import com.umc.ttt.domain.bookLetter.bookLetterRepository.BookLetterBookRepository;
+import com.umc.ttt.domain.bookLetter.repository.BookLetterBookRepository;
 import com.umc.ttt.domain.bookLetter.entity.BookLetterBook;
-import com.umc.ttt.domain.bookLetter.handler.BookLetterBookHandler;
+import com.umc.ttt.global.apiPayload.exception.handler.BookLetterBookHandler;
 import com.umc.ttt.domain.member.entity.Member;
 import com.umc.ttt.global.apiPayload.code.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
