+++ src/main/java/com/umc/ttt/domain/bookLetter/repository/BookLetterBookRepository.java
@@ -1,10 +1,7 @@
-package com.umc.ttt.domain.bookLetter.bookLetterRepository;
+package com.umc.ttt.domain.bookLetter.repository;
 
-import com.umc.ttt.domain.bookLetter.entity.BookLetter;
 import com.umc.ttt.domain.bookLetter.entity.BookLetterBook;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface BookLetterBookRepository extends JpaRepository<BookLetterBook,Long> {
 }
