+++ src/main/java/com/umc/ttt/domain/bookLetter/service/BookLetterCommandService.java
@@ -3,7 +3,7 @@ package com.umc.ttt.domain.bookLetter.service;
 import com.umc.ttt.domain.bookLetter.dto.BookLetterRequestDTO;
 import com.umc.ttt.domain.bookLetter.entity.BookLetter;
 import com.umc.ttt.domain.member.entity.Member;
-import com.umc.ttt.home.dto.HomeResponseDTO;
+import com.umc.ttt.domain.home.dto.HomeResponseDTO;
 import org.springframework.data.domain.Page;
 import org.springframework.web.multipart.MultipartFile;
