+++ src/main/java/com/umc/ttt/domain/bookLetter/service/BookLetterCommandServiceImpl.java
@@ -6,15 +6,15 @@ import com.umc.ttt.domain.book.entity.BookFormatCategory;
 import com.umc.ttt.domain.book.repository.BookCategoryRepository;
 import com.umc.ttt.domain.book.repository.BookFormatCategoryRepository;
 import com.umc.ttt.domain.book.repository.BookRepository;
-import com.umc.ttt.domain.bookLetter.Converter.BookLetterConverter;
-import com.umc.ttt.domain.bookLetter.bookLetterRepository.BookLetterBookRepository;
-import com.umc.ttt.domain.bookLetter.bookLetterRepository.BookLetterCategoryRepository;
-import com.umc.ttt.domain.bookLetter.bookLetterRepository.BookLetterRepository;
+import com.umc.ttt.domain.bookLetter.converter.BookLetterConverter;
+import com.umc.ttt.domain.bookLetter.repository.BookLetterBookRepository;
+import com.umc.ttt.domain.bookLetter.repository.BookLetterCategoryRepository;
+import com.umc.ttt.domain.bookLetter.repository.BookLetterRepository;
 import com.umc.ttt.domain.bookLetter.dto.BookLetterRequestDTO;
 import com.umc.ttt.domain.bookLetter.entity.BookLetter;
 import com.umc.ttt.domain.bookLetter.entity.BookLetterBook;
 import com.umc.ttt.domain.bookLetter.entity.BookLetterCategory;
-import com.umc.ttt.domain.bookLetter.handler.BookLetterCategoryHandler;
+import com.umc.ttt.global.apiPayload.exception.handler.BookLetterCategoryHandler;
 import com.umc.ttt.domain.member.entity.Member;
 import com.umc.ttt.domain.member.repository.MemberPreferredCategoryRepository;
 import com.umc.ttt.global.apiPayload.code.status.ErrorStatus;
@@ -23,8 +23,8 @@ import com.umc.ttt.global.apiPayload.exception.handler.BookLetterHandler;
 import com.umc.ttt.global.aws.s3.entity.Uuid;
 import com.umc.ttt.global.aws.s3.repository.UuidRepository;
 import com.umc.ttt.global.aws.s3.service.AmazonS3Manager;
-import com.umc.ttt.home.converter.HomeConverter;
-import com.umc.ttt.home.dto.HomeResponseDTO;
+import com.umc.ttt.domain.home.converter.HomeConverter;
+import com.umc.ttt.domain.home.dto.HomeResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
