+++ src/main/java/com/umc/ttt/domain/home/service/HomeService.java
@@ -1,7 +1,7 @@
-package com.umc.ttt.home.service;
+package com.umc.ttt.domain.home.service;
 
+import com.umc.ttt.domain.home.dto.HomeResponseDTO;
 import com.umc.ttt.domain.member.entity.Member;
-import com.umc.ttt.home.dto.HomeResponseDTO;
 
 public interface HomeService {
     HomeResponseDTO.viewHomeResultDTO getHomeData(Member member);
