+++ src/main/java/com/umc/ttt/domain/home/service/HomeServiceImpl.java
@@ -1,12 +1,11 @@
-package com.umc.ttt.home.service;
+package com.umc.ttt.domain.home.service;
 
 import com.umc.ttt.domain.bookClub.service.BookClubQueryService;
-import com.umc.ttt.domain.bookLetter.entity.BookLetter;
 import com.umc.ttt.domain.bookLetter.service.BookLetterCommandService;
+import com.umc.ttt.domain.home.converter.HomeConverter;
+import com.umc.ttt.domain.home.dto.HomeResponseDTO;
 import com.umc.ttt.domain.member.entity.Member;
 import com.umc.ttt.domain.review.service.ReviewCommandService;
-import com.umc.ttt.home.converter.HomeConverter;
-import com.umc.ttt.home.dto.HomeResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
