--- src/main/java/com/umc/ttt/domain/member/handler/MemberHandler.java
@@ -1,10 +1,0 @@
-package com.umc.ttt.domain.member.handler;
-
-import com.umc.ttt.global.apiPayload.code.BaseErrorCode;
-import com.umc.ttt.global.apiPayload.exception.GeneralException;
-
-public class MemberHandler extends GeneralException {
-    public MemberHandler(BaseErrorCode errorCode) {
-        super(errorCode);
-    }
-}
