+++ src/main/java/com/umc/ttt/global/apiPayload/exception/handler/BookLetterBookHandler.java
@@ -1,4 +1,4 @@
-package com.umc.ttt.domain.bookLetter.handler;
+package com.umc.ttt.global.apiPayload.exception.handler;
 
 import com.umc.ttt.global.apiPayload.code.status.ErrorStatus;
 import com.umc.ttt.global.apiPayload.exception.GeneralException;
